Code refactoring

+++ 코딩테스트 연습문제/Lv. 1/부족한 금액 계산하기.java
@@ -1,9 +1,9 @@
 class Solution {
     public long solution(int price, int money, int count) {
         long total = 0;
-        
+
         for (int i = 1; i <= count; i++) {
-            total += (price * i);
+            total += ((long)price * i);
         }
         return (total > money) ? (total - money) : 0;
     }
